requirements.txt
cryptography
csbdeep
elasticdeform
Flask
Flask-Dance
Flask-Login
Flask-Session
Flask-SocketIO
Flask-SQLAlchemy
matplotlib
numpy
oauthlib
opencv-python
Pillow
psutil
PyJWT
python-dotenv
requests
scikit-image
scipy
SQLAlchemy
stardist
torch
tqdm
waitress
Werkzeug
